0.3.5 Bot - Refactoring, Profile & Change Profile
