+++ .github/workflows/unittest_telebot.yml
@@ -16,5 +16,5 @@ jobs:
         python-version: 3.x
 
     - name: Run unit tests
-      run: python -m unittest telegramhelper_test.py
+      run: python -m unittest botfunctions/functions_unittesting.py
       working-directory: Telegram Bot
